requirements.txt
fastapi
uvicorn
httpx
python-dotenv
python-telegram-bot
supabase
openai
google-generativeai